Extract thumbnail helper methods to new module
