+++ app/helpers/application_helper.rb
@@ -72,19 +72,6 @@ module ApplicationHelper
     "#{request.protocol}#{request.host_with_port}/catalog?#{query}"
   end
 
-  def thumbnail_from_manifest(document, image_options = { height: '400', width: '400' })
-    image_tag "#{document['thumbnail_base']}/full/!#{image_options[:height]},#{image_options[:width]}/0/default.png"
-  end
-
-  def render_square_thumbnail(document)
-    image_tag "#{document['thumbnail_base']}/full/!400,400/0/default.png"
-  end
-
-  def render_fullsize_thumbnail(document)
-    return unless document.nil?
-    image_tag "#{document['thumbnail_base']}/full/!600,600/0/default.jpg"
-  end
-
   def doc_path(document)
     "/catalog/#{document['id']}"
   end
